requirements.txt
PyGithub==1.58.1
matplotlib==3.7.1